Split out errors and installation
