+++ src/registry.js
@@ -7,119 +7,18 @@ import {publishReplay} from 'rxjs/operator/publishReplay'
 import {httpGetJSON} from './util'
 import * as config from './config'
 import * as imCache from './im_cache'
+import {PackageRootError, VersionError} from './errors'
 
 /**
- * error class for the case when the required version target is not available
- * (= the package is in the registry, but the available version is not
- * available).
- * @extends Error
+ * Validate the given body.
+ * @param  {String} url  - url from which the document has been retrieved.
+ * @param  {Object} body - package root as JSON object.
+ * @throws {PackageRootError}
+ * @throws {Error}
  */
-export class VersionError extends Error {
-  /**
-   * create instance.
-   * @param {String} name - name of the package.
-   * @param {String} version - unavailable version number.
-   * @param {Array.<String>} - array of available version numbers.
-   * @param {Array.<String>} options.available - an array of available versions
-   */
-  constructor (name, version, available) {
-    super(`no satisying version of ${name} in ${available.join(', ')} for ${version}`)
-    this.name = 'VersionError'
-    this.pkgName = name
-    this.version = version
-    this.available = available
-  }
-
-  /**
-   * name of the error (not the package name).
-   *
-   * @name VersionError#name
-   * @type String
-   * @default "VersionError"
-   * @readonly
-   */
-
-  /**
-   * package name.
-   *
-   * @name VersionError#pkgName
-   * @type String
-   * @readonly
-   */
-
-  /**
-   * target version that could not be found.
-   *
-   * @name VersionError#version
-   * @type String
-   * @readonly
-   */
-
-  /**
-   * all available version numbers of the supplied package.
-   *
-   * @name available
-   * @type Array.<String>
-   * @readonly
-   */
-}
-
-/**
- * error class for the case when the package root is not "valid", meaning the
- * document that we got from the registry does not represent a valid package
- * root document.
- */
-export class PackageRootError extends Error {
-  /**
-   * create instance.
-   * @param  {String} url - url of the package root.
-   * @param  {Object} body - body of the invalid package root.
-   */
-  constructor (url, body) {
-    if (body && body.error) {
-      super(`invalid package root at ${url}: ${body.error}`)
-    } else {
-      super(`invalid package root at ${url}`)
-    }
-    this.name = 'PackageRootError'
-    this.url = url
-    this.body = body
-  }
-
-  /**
-   * name of the error.
-   *
-   * @name PackageRootError#name
-   * @type String
-   * @default "PackageRootError"
-   * @readonly
-   */
-
-  /**
-   * url of the package root.
-   * @name PackageRootError#url
-   * @type String
-   * @readonly
-   */
-
-  /**
-   * body of the invalid package root.
-   * @name PackageRootError#body
-   * @type Object
-   * @readonly
-   */
-
-  /**
-   * Validate the given body.
-   * @param  {String} url  - url from which the document has been retrieved.
-   * @param  {Object} body - package root as JSON object.
-   * @throws {PackageRootError}
-   * @throws {Error}
-   */
-  static validate (url, body) {
-    if (!body || body.error || typeof body.versions !== 'object') {
-      throw new PackageRootError(url, body)
-    }
+export function validatePackageRoot (uri, body) {
+  if (!body || body.error || typeof body.versions !== 'object') {
+    throw new PackageRootError(uri, body)
   }
 }
 
@@ -135,7 +34,7 @@ export function httpGetPackageRoot (name) {
   if (cached) return cached
   const result = httpGetJSON(uri)
     ::retry(5)
-    ::_do((body) => PackageRootError.validate(uri, body))
+    ::_do((body) => validatePackageRoot(uri, body))
     ::publishReplay().refCount()
   return imCache.set(uri, result)
 }
